# Conflicts:
#	src/test/java/eu/europa/eudi/elements/android/WalletElements.java
#	src/test/java/eu/europa/eudi/elements/ios/WalletElements.java
#	src/test/java/eu/europa/eudi/stepdefs/AutomatedStepDefs.java

+++ src/test/java/eu/europa/eudi/utils/factory/MobileDeviceLogger.java
@@ -11,7 +11,7 @@ public class MobileDeviceLogger {
     private static Process logcatProcess;
     private static Thread logcatThread;
 
-    public static void startLogging(String featureDirPath,
+    public static File startLogging(String featureDirPath,
                                     String featureName,
                                     String scenarioNumber,
                                     String platform) {
@@ -39,6 +39,10 @@ public class MobileDeviceLogger {
                     featureName + "_" + scenarioNumber + ".log"
             );
 
+            if (!logDir.exists() && !logDir.mkdirs()) {
+                throw new IOException("Cannot create log directory: " + logDir);
+            }
+
             if (logFile.exists() && !logFile.delete()) {
                 throw new IOException("Cannot delete old log file: " + logFile);
             }
@@ -57,7 +61,7 @@ public class MobileDeviceLogger {
             // ----------------------------
             if ("browserstack".equalsIgnoreCase(env)) {
                 System.out.println("BrowserStack mode → skipping local logcat, file created");
-                return;
+                return logFile;
             }
 
             // ----------------------------
@@ -111,8 +115,11 @@ public class MobileDeviceLogger {
             logcatThread.setDaemon(true);
             logcatThread.start();
 
+            return logFile;
+
         } catch (Exception e) {
             e.printStackTrace();
+            return null;
         }
     }
